+++ src/as_display/debug/mod.rs
@@ -1,7 +1,0 @@
-mod as_debug_wrapper;
-mod debug_as_display;
-mod debug_wrapper;
-
-pub use as_debug_wrapper::AsDebugWrapper;
-pub use debug_as_display::DebugAsDisplay;
-pub use debug_wrapper::DebugWrapper;