+++ src/as_display/display/mod.rs
@@ -1,7 +1,0 @@
-mod as_display_wrapper;
-mod display_as_debug;
-mod display_wrapper;
-
-pub use as_display_wrapper::AsDisplayWrapper;
-pub use display_as_debug::DisplayAsDebug;
-pub use display_wrapper::DisplayWrapper;